refactor(client/src): team review
