+++ packages/client/src/store/store.ts
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 
 import { appReducer } from './features/app/appSlice';
 import { authReducer } from './features/auth/authSlice';
-import { uiReducer } from './features/ui/ui-slice';
+import { uiReducer } from './features/ui/uiSlice';
 
 export const store = configureStore({
   reducer: {
